requirements.txt
python-dotenv==0.19.2
discord.py==1.7.3
gspread==5.1.1
json